+++ src/games/domino/players/strategies/utils/__init__.py
@@ -1,0 +1,2 @@
+from .utils import *
+from .game  import *
