+++ src/games/domino/players/strategies/utils/game.py
@@ -1,22 +1,6 @@
 import random
 
 
-def is_end(history, pieces_per_player):
-    heads = []
-    player_pieces = {}
-    first_move = True
-    for e, *d in history:
-        if e.name == 'MOVE':
-            player, piece, head = d
-            if first_move:
-                heads = list(piece)
-                first_move = False
-            else:
-                heads[head] = piece[piece[0] == heads[head]]
-            player_pieces[player] = player_pieces.get(player, 0) + 1
-    return any([pieces_per_player - x <= 2 for x in player_pieces.values()])
-
-
 class EdmondsKarp:
     class edge:
         def __init__(self, start, end, f, cap):
@@ -149,3 +133,6 @@ def game_hand_builder(pieces, missing, remaining, number_of_pieces=7):
             pieces[edge.start - 1].append(remaining[edge.end - 5])
     
     return pieces
+
+
+__all__ = [game_data_collector, game_hand_builder, remaining_pieces]
